Upgrade to v0.0.3 test framework
